pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "binance-terminal"
version = "1.0.0"
description = "A professional trading application for Binance cryptocurrency exchange"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["binance", "trading", "cryptocurrency", "gui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "python-binance==1.0.19",
    "PySide6>=6.5.0,<7.0.0",
    "matplotlib>=3.7.0,<4.0.0",
    "mplfinance>=0.12.9",
    "pandas>=2.0.0,<3.0.0",
    "requests>=2.28.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "websocket-client>=1.6.0,<2.0.0",
    "numpy>=1.21.0,<2.0.0",
    "aiohttp>=3.8.0,<4.0.0",
    "cryptography>=41.0.0,<42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "pre-commit>=2.15.0",
    "pyinstaller==6.11.0",
    "ruff>=0.1.0",
]

build = [
    "pyinstaller==6.11.0",
]

[project.urls]
Homepage = "https://github.com/AhmetNA/binance-terminal"
Repository = "https://github.com/AhmetNA/binance-terminal.git"
Documentation = "https://github.com/AhmetNA/binance-terminal/docs"
"Bug Tracker" = "https://github.com/AhmetNA/binance-terminal/issues"

[project.scripts]
binance-terminal = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
